check if poll module is enabled before use the collection of it

+++ app/code/community/IntegerNet/MagentoLocalized/data/magento_localized_setup/data-install-0.1.0.php
@@ -12,11 +12,14 @@
 /* @var $installer Mage_Core_Model_Resource_Setup */
 $installer = $this;
 
-/** @var $pollCollection Mage_Poll_Model_Resource_Poll_Collection */
-$pollCollection = Mage::getModel('poll/poll')->getCollection();
-foreach($pollCollection as $poll) {
-    /** @var $poll Mage_Poll_Model_Poll */
-    $poll->delete();
+if(Mage::helper('core')->isModuleEnabled('poll')) {
+    /** @var $pollCollection Mage_Poll_Model_Resource_Poll_Collection */
+    $pollCollection = Mage::getModel('poll/poll')->getCollection();
+    foreach($pollCollection as $poll) {
+        /** @var $poll Mage_Poll_Model_Poll */
+        $poll->delete();
+    }
 }
+
 
 $installer->endSetup();
